Moving to dangerous grid world
