+++ src/envs/__init__.py
@@ -2,4 +2,4 @@ from src.envs.frozen_lake.utils import *
 from src.envs.CMDP import CMDP, LagrangianMDP, LagrangianMDPMonitor
 from src.envs.frozen_lake.frozen_lake_custom import *
 from src.envs.dummy_envs import *
-from src.envs.cliff_walking.cliff_walking_custom import CliffWalkingEnvCustom
+from src.envs.dangerous_grid_world.dangerous_grid_world_custom import DangerousGridWorldEnvCustom
